Don't depend on jupyter (#61)
`jupyter` is a conda metapackage for enduser to conviently install all `jupyter` components and should not be used as `run` dependency in a package. The dependencies on `notebook` and `ipython` should be enough for this to run.

Background on why this is important: `jupyter` includes a dependency on `qtconsole` which in turn pulls in `qt` that adds a lot of size to a conda installation.